chapter10/pyproject.toml
[project]
name = "mcp-agent"
version = "0.1.0"
description = "LLM-integrated MCP Agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.0",
    "fastmcp>=0.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.0.0",
    "pyyaml>=6.0.0",
]